backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.43
pymysql==1.1.0
alembic==1.13.0
email-validator==2.2.0